[Silver III] Title: 주유소, Time: 328 ms, Memory: 39340 KB, Score: 100 point -BaekjoonHub

+++ Java/백준/Silver/13305. 주유소/주유소.java
@@ -9,20 +9,20 @@ public class Main {
         String[] meter= br.readLine().split(" ");
         String[] cost= br.readLine().split(" ");
 
-        int[] distance= new int[n-1];
-        int[] costs= new int[n-1];
+        long[] distance= new long[n-1];
+        long[] costs= new long[n-1];
 
 
         // 마지막 도시는 어짜피 도착지니까 무시
-        int ans=0; // 총 가격
+        long ans=0; // 총 가격
 
         for( int i=0; i<n-1; i++) {
-            distance[i]= Integer.parseInt(meter[i]);
-            costs[i]= Integer.parseInt(cost[i]);
+            distance[i]= Long.parseLong(meter[i]);
+            costs[i]= Long.parseLong(cost[i]);
         }
 
         // 최소 cost로 계속 구매
-        int min= costs[0];
+        long min= costs[0];
         for(int i=0; i<n-1; i++){
                 if(costs[i] <= min) {
                     min= costs[i];
@@ -30,7 +30,7 @@ public class Main {
                 ans += min * distance[i];
             }
 
-        bw.write(Integer.toString(ans));
+        bw.write(Long.toString(ans));
         bw.flush();
         bw.close();
         br.close();
